Fix the 'it is a draw' message in getScore func
Previously if the user had won a game, win would be set to true but
unless the user lost the next round, win would still be true and the
players score would go up by one even if they drew with the computer.

Adding the lose variable and initializing both win and lose with falsy
values prevents this from happening, displaying the proper draw message
when it is a draw.