Create a new function to display the winner/loser
Before the winner & loser was displayed inside the play game function
when 5 games were played, however this clutters the main purpose of the
playGame function which is to play one game and increment the gamePlayed
counter.

In order to clean up the code i've created a new function to display
the winner/loser message and create a reset button once five games have
been played.